Move to index and change channels, ready for merge
