+++ index.js
@@ -3,6 +3,7 @@ const fs = require("fs")
 const fetch = require("node-fetch")
 const Discord = require("discord.js")
 const client = new Discord.Client()
+const { registerFont, createCanvas, loadImage } = require("canvas")
 require("dotenv").config()
 
 //Import data, assets and commands
@@ -25,7 +26,6 @@ const notAllowed = "732298639736570007" //vote_no emoji
 const stats = require("./events/stats.js")
 const inactives = require("./events/inactives.js")
 const unzalgo = require("./events/unzalgo.js")
-const guildMemberAdd = require("./events/guildMemberAdd.js")
 
 
 //Run when bot is ready
@@ -324,12 +324,71 @@ client.on("messageReactionRemove", async (reaction, user) => {
 })
 
 //Run when someone joins
-client.on("guildMemberAdd", member => guildMemberAdd.execute(member))
+client.on("guildMemberAdd", member => {
+
+  //Define assets and create canvas
+  registerFont("./assets/Bitter-Regular.ttf", { family: "Bitter" })
+  registerFont("./assets/Bitter-Bold.ttf", { family: "Bitter-Bold" })
+  const canvas = createCanvas(800, 200)
+  const ctx = canvas.getContext("2d")
+  const userName = member.user.username
+  const userAvatar = member.user.displayAvatarURL({ format: "png" })
+  const memberCount = `${member.guild.memberCount}`
+
+  loadImage("./assets/joinBackground.png").then(bg => {
+    let nameWidth
+
+    //GENERAL
+    //Add background and set basic styling
+    ctx.drawImage(bg, 0, 0, 800, 200)
+    ctx.fillStyle = "white"
+
+
+    //TEXT
+    //Measure text widths
+    ctx.font = "37.5px Bitter"
+    let welcome = ctx.measureText("Welcome ")
+    ctx.font = "37.5px Bitter-Bold"
+    let name = ctx.measureText(userName)
+    if (name.width > (550 - welcome.width)) nameWidth = (550 - welcome.width)
+    else nameWidth = name.width
+
+    //Draw 'Welcome ' and '!'
+    ctx.font = "37.5px Bitter"
+    ctx.fillText("Welcome ", 200, 92.5)
+    ctx.fillText("!", (200 + welcome.width + nameWidth), 92.5)
+
+    //Draw username
+    ctx.font = "37.5px Bitter-Bold"
+    ctx.fillText(userName, (200 + welcome.width), 92.5, (550 - welcome.width))
+
+    //Draw member count
+    ctx.font = "30px Bitter"
+    ctx.fillText("You're member #" + memberCount, 200, 132.5)
+
+    //ICON
+    //Draw a circle for the image to go into
+    ctx.beginPath()
+    ctx.arc(100, 100, 75, 0, 2 * Math.PI)
+    ctx.closePath()
+
+    //Put the image in the circle
+    loadImage(userAvatar).then((userPic) => {
+      ctx.clip()
+      ctx.drawImage(userPic, 25, 25, 150, 150)
+      ctx.restore()
+
+      //OUTPUT
+      const attachment = new Discord.MessageAttachment(canvas.toBuffer(), "join.png")
+      member.guild.channels.cache.get("549882021934137354").send(`<@${member.user.id}> just joined! Welcome! 🎉!`, attachment) //join-leave
+    })
+  })
+})
 
 //Run when someone leaves
 client.on("guildMemberRemove", member => {
   //Leave message
-  member.guild.channels.cache.get("730042612647723058").send(`**${member.user.tag}** left the server 🙁`) //join-leave
+  member.guild.channels.cache.get("549882021934137354").send(`**${member.user.tag}** just left the server 🙁`) //join-leave
 
   //Run if the member who leaves had the Bot Translator/Proofreader/Manager roles
   const botRole = member.roles.cache.find(role => role.name.startsWith("Bot ") && role.id !== "732615152246980628")
